sky box, area pattern
