+++ core/frontend/src/render/MockRender.ts
@@ -25,11 +25,12 @@ import { PointCloudArgs } from "../common/internal/render/PointCloudPrimitive";
 import { GraphicList, RenderGraphic } from "./RenderGraphic";
 import { RenderMemory } from "./RenderMemory";
 import { RenderPlan } from "../internal/render/RenderPlan";
-import { RenderAreaPattern, RenderSystem } from "./RenderSystem";
+import { RenderSystem } from "./RenderSystem";
 import { RenderTarget } from "./RenderTarget";
 import { Scene } from "./Scene";
 import { _implementationProhibited } from "../common/internal/Symbols";
 import { RenderGeometry } from "../internal/render/RenderGeometry";
+import { RenderAreaPattern } from "../internal/render/RenderAreaPattern";
 
 /** Contains extensible mock implementations of the various components of a RenderSystem, intended for use in tests.
  * Use these for tests instead of the default RenderSystem wherever possible because:
